Moved dialogs to their own folder
