+++ ui/lib/src/dialogs/where-svg-marker-dialog.ts
